requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
pillow>=9.0.0
numpy>=1.23.0
python-dateutil>=2.8.0
